Update Chrome.ahk
When the profilepath given by the user does not exist, try to create it.

In this way, chrome can be simply run with a brand-new identity.

Therefore, the user can specify an existing identity to run, or simply create a new identity for automation and keep it there all the time.